### 2018-12-14 #1129 @whedon
[PRE REVIEW]: strandCheckR: An R package for quantifying and removing double strand sequences for strand-specific RNA-seq
**Submitting author:** @tothuhien (<a href="http://orcid.org/0000-0003-2902-3655">Thu-Hien To</a>)
**Repository:** <a href="https://github.com/UofABioinformaticsHub/strandCheckR" target ="_blank">https://github.com/UofABioinformaticsHub/strandCheckR</a>
**Version:** 1.0.0
