### 2018-12-14 #1129 @tothuhien
Thanks Arfon! I've just changed the YAML header following the example. Hope
it works now.
Hien

On Fri, Dec 14, 2018 at 8:28 PM Arfon Smith <notifications@github.com>
wrote:

> Hi @tothuhien <https://github.com/tothuhien> - please update the YAML
> header of your paper.md to exactly match the structure of this example:
> https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
>
