### 2018-12-14 #1129 @tothuhien
Thanks for building it. I've checked the article proof and it's all good!
Hien
