### 2018-12-28 #1145 @whedon
[REVIEW]: strandCheckR: An R package for quantifying and removing double strand sequences for strand-specific RNA-seq
**Submitting author:** @tothuhien (<a href="http://orcid.org/0000-0003-2902-3655">Thu-Hien To</a>)
**Repository:** <a href="https://github.com/UofABioinformaticsHub/strandCheckR" target ="_blank">https://github.com/UofABioinformaticsHub/strandCheckR</a>
**Version:** v1.1.2

  - No [license](https://github.com/UofABioinformaticsHub/strandCheckR/issues/2) file. It is licensed according to CRAN and Bioconductor standards, but not commonly accepted OSS project standards.
  - This was resolved by [this commit](https://github.com/UofABioinformaticsHub/strandCheckR/commit/1340208624961bf0a5ee142cb539b3d5099c08a9)
  - Yes, based on the git contributions, it seems the authors list is appropriate and complete.

  - Yes, installation proceeds without problems. As noted in [this issue](https://github.com/UofABioinformaticsHub/strandCheckR/issues/6), I do suggest removing the list of required packages from the README, as Bioconductor installs the dependencies without difficulty.
  - Yes, I have confirmed that I can run the code provided in the vignette, and that it behaves as outlined in the vignette and paper.
  - The paper has a performance claim that is not verified. See [this issue](https://github.com/UofABioinformaticsHub/strandCheckR/issues/8).

  - This is stated in the paper, and in the package documentation, but not in the README. Copying over the statement from the DESCRIPTION should be sufficient. See [this issue](https://github.com/UofABioinformaticsHub/strandCheckR/issues/4).
  - Dependencies are handled just fine by the Bioconductor installer.
  - There is no example of usage in the README, but the included vignette does provide a fully worked example. I think including the commands listed in the package level documentation into the README would be sufficient. See [this issue](https://github.com/UofABioinformaticsHub/strandCheckR/issues/5).
  - The core functionality is documented, in that there are descriptions for each function and their arguments.
  - However, some of the documentation lacks examples, and there are places where I believe the documentation could be improved, as outlined in [this pull request](https://github.com/UofABioinformaticsHub/strandCheckR/pull/3) and [this issue](https://github.com/UofABioinformaticsHub/strandCheckR/issues/9).
  - There are automated tests.
  - No Contributing file or description provided. See [this issue](https://github.com/UofABioinformaticsHub/strandCheckR/issues/7)
