### 2019-01-02 UofABioinformaticsHub/strandCheckR#2 @rmflight
JOSS Review: no license file
Although the DESCRIPTION file lists the GPL (>= 2) license, the project is missing a proper LICENSE file with the text of the license.

Although CRAN & Bioconductor forbids including the full text of licenses that are already covered by CRAN approved licenses, one can use the `.Rbuildignore` file to keep the license file from being included in the built and checked package.
