### 2019-01-04 UofABioinformaticsHub/strandCheckR#2 @tothuhien
Thanks @rmflight for reviewing the package. @steveped has added the LICENSE file as you suggested. Please let us know if there's any other issue.
Hien
