### 2019-01-04 UofABioinformaticsHub/strandCheckR#2 @rmflight
I'm working through the rest of the review now @tothuhien, and hope to have it done early next week.
