### 2019-01-04 UofABioinformaticsHub/strandCheckR#4 @rmflight
JOSS Review: README statement of need
Include an actual statement of need, describing why the package is useful. This could be taken from the DESCRIPTION file.
