### 2019-01-04 UofABioinformaticsHub/strandCheckR#5 @rmflight
JOSS Review: README example usage
Include some example usage in the README.

Does not need to be as extensive as the vignette or the paper, but showing some basic functionality, including the output of commands (I would use a README.Rmd to generate README.md), the commands in the package level documentation should be sufficient.
