### 2019-01-04 UofABioinformaticsHub/strandCheckR#6 @rmflight
JOSS Review: README remove dependency installation
Remove installing dependent packages in the README installation instructions (currently line 14 in README.md). Given that the package is on Bioconductor, I would include both the Bioconductor instructions and the `devtools::install_github` instructions. I used the `devtools::install_github` instructions, and all the dependencies were properly installed on a clean R installation, so the dependency installation line is no longer required.
