### 2019-01-04 UofABioinformaticsHub/strandCheckR#7 @rmflight
JOSS Review: Community Guidelines
There is no `Contributing.md` file, or a statement regarding contributions in the README.md, or anything regarding reporting issues or submitting pull requests.
