### 2019-01-04 UofABioinformaticsHub/strandCheckR#8 @rmflight
JOSS Review: Performance Claim
The primary performance claim is in this statement in their paper: *For a human RNA-seq bam file of ~4Gb, it takes less than 3 minutes to scan the whole file and return a Data Frame of about 1.5Gb*. I do not have access to this size of file, so I cannot check this claim. However, given that the software is making use of the `Rsamtools` and `GenomicRanges` packages for the quantification, I have no reason to question these claims. The authors might consider qualifying the claim by including information about the compute environment (CPU, RAM) that generated this level of performance.
