### 2019-01-04 UofABioinformaticsHub/strandCheckR#9 @rmflight
JOSS Review: Functionality Documentation
- [x] The `sequences` parameter in `getWinFromBamFile` is not really described. I haven't looked at BAM/SAM files very much, but I can't figure out what this actually corresponds to. I see that it is used to subset the sequences from the BAM file using `seqinfo()`, but I can't get `seqinfo` to work on after reading in the example file using `scanBam`. If it corresponds to subsetting sequences on a **chromosome** or some other identifier, that should be included in the documentation. From the example, it looks like this results in **Seq** being "chr10" in the `win` object.
- [x] The various columns in the data frame from `getWinFromBamFile` are not documented. As noted above, it **appears** that `sequences` in `getWinFromBamFile` corresponds to `Seq` in the output data.frame, but this is not made explicit anywhere. What information is actually contained in many other columns is not immediately intuitive either. This information could be in `Details`.
- [x] Inconsistent parameters among `getWinFrom*` functions. `readProp` is a parameter in `*FromBamFile` and `*FromIRanges`, but not `*FromGranges`, with no explanation why it's not needed.
- [x] Inconsistent naming between `getWinFromIRanges` and `getWinFromGranges`, would have expected `getWinFromGRanges`.
- [x] Inconsistent defaults for `winWidth` and `winStep`, in some documentation `1000` / `100` is shown, in others `1000L` / `100L` is shown. One is a double, one is an integer.
- [x] In `plotHist`, the `normalize_by` variable seems not well explained, and there are no further details in the document, nor any examples of use. At the very least, it might be modified from *the column names of windows that will be used to normalize the read count or read coverage into proportion* to *the columns that will be used to normalize the read count or read coverage into a proportion*.
- [x] Inconsistent use of "column names" (plotHist) vs "colnames" (plotWin). Would suggest using "column names" everywhere.
- [x] Some inconsistent argument naming, namely `normalize_by` and `group_by`, should these be `normalizeBy` and `groupBy`??
