### 2019-01-07 UofABioinformaticsHub/strandCheckR#6 @smped
Hi Robert,
Thanks for going through everything so thoroughly. This has now been changed as of commit `c883c26808fe7234b6c6da60d5077a1d623c70f0` so hopefully that's another one to tick off the list.
Cheers,
Steve
