### 2019-01-09 UofABioinformaticsHub/strandCheckR#4 @smped
This is now up to date as of commit 87517b1
