### 2019-01-09 UofABioinformaticsHub/strandCheckR#5 @smped
This is also done as of commit `baf6706`
