### 2019-01-09 UofABioinformaticsHub/strandCheckR#7 @smped
As of commit `025581c` a `BugReports` line has been added to the `DESCRIPTION` file, and information regarding author contributions and support have also been added to `README.md`
