### 2019-01-09 UofABioinformaticsHub/strandCheckR#7 @rmflight
I think that satisfies the requirements for the review, and will close this issue.

I would encourage a bit more formality, and the use of a CONTRIBUTING or CONDUCT file that lays out the basic terms, including what kinds of contributions result in the person being added to the author list of the package.

See this [github post](https://help.github.com/articles/setting-guidelines-for-repository-contributors/), and an example on [other](https://github.com/rmflight/knitrProgressBar/blob/master/CONDUCT.md)
