### 2019-01-14 #1145 @mgymrek
@rmflight great. can you paste as a comment and recommended changes especially regarding items from the checklist you did not find satisfactory?
