### 2019-01-14 #1145 @rmflight
Hi @mgymrek . I actually created issues for all of this in [the repo](https://github.com/UofABioinformaticsHub/strandCheckR/issues?q=is%3Aissue+author%3Armflight), and have been checking items from the main review above as they have been addressed. I can put all the comments on here, but didn't think that was necessary.
