### 2019-01-17 UofABioinformaticsHub/strandCheckR#9 @tothuhien
Hi Robert, I think we've addressed all those points now. Note that I've changed a bit some function names to distinguish some functions that sounded similar but did different things (i.e. getWinFromIRanges and getWinFromGRange).
Cheers,
Hien
