### 2019-01-21 #1145 @mgymrek
@rmflight that's fine. Can you update this comment thread when all of those issues have been addressed?
