### 2019-01-23 UofABioinformaticsHub/strandCheckR#9 @rmflight
Looks like everything is addressed. Thanks @tothuhien !
