### 2019-01-24 UofABioinformaticsHub/strandCheckR#8 @tothuhien
I couldn't find the file of 4Gb that I've tested before, so I've used another human RNA-seq bam file and see the same performance. I've modified that in the paper and added the information about the computer I've used.
Cheers,
Hien
