### 2019-02-01 #1145 @steveped
Hi @mgymrek,

I've just created a release with the DOI & added the badge to the README. It's 10.5281/zenodo.2555032. Hopefully I got the process right for a first timer. :)
