### 2019-02-15 #1145 @mgymrek
@wedon set v1.1.2 as version
