### 2019-02-17 #1145 @arfon
@mgymrek - in the future, please ping `@openjournals/joss-eics` when you're ready to accept a paper. This will then notify the managing editor on rotation.
